requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas>=2.1.0
numpy>=1.25.0
scikit-learn>=1.3.0
joblib>=1.3.2
psycopg2-binary>=2.9.7
SQLAlchemy>=2.0.41
Flask-SQLAlchemy>=3.0.5
Flask-Migrate>=4.0.5
python-dotenv>=1.0.0
Werkzeug>=2.3.7
alembic>=1.12.0
requests>=2.31.0
